read in keystore file
