+++ anchor/keygen/Cargo.toml
@@ -6,3 +6,5 @@ authors = ["Sigma Prime <contact@sigmaprime.io>"]
 
 [dependencies]
 clap = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
