+++ anchor/keygen/src/keystore.rs
@@ -1,0 +1,63 @@
+use serde::{Deserialize, Serialize};
+use std::fs::File;
+use std::io::BufReader;
+
+// Representation of a keystore file generated by the deposit cli
+//https://github.com/ssvlabs/ssv-keys/blob/df1f0a5edd7ce95ad5ab0298c58e83908216557b/src/lib/EthereumKeyStore/EthereumKeyStore.ts#L7
+#[derive(Serialize, Deserialize, Debug)]
+struct Keystore {
+    crypto: Crypto,
+    description: String,
+    pubkey: String,
+    path: String,
+    uuid: String,
+    version: u32,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct Crypto {
+    kdf: Kdf,
+    checksum: Checksum,
+    cipher: Cipher,
+}
+#[derive(Serialize, Deserialize, Debug)]
+struct Checksum {
+    function: String,
+    //params // the params
+    message: String,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct Cipher {
+    function: String,
+    params: CipherParams,
+    message: String,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct CipherParams {
+    iv: String,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct Kdf {
+    function: String,
+    params: KdfParams,
+    message: String,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct KdfParams {
+    dklen: u32,
+    n: u32,
+    r: u32,
+    p: u32,
+    salt: String,
+}
+
+
+pub fn parse_keystore(keystore: File) {
+    let reader = BufReader::new(keystore);
+    let data: Keystore = serde_json::from_reader(reader).unwrap();
+    println!("{:?}", data);
+}
