+++ anchor/keygen/src/lib.rs
@@ -1,40 +1,8 @@
+use cli::SharedKeygenOptions;
 pub use cli::{Keygen, KeygenSubcommands, Manual, Onchain};
+use std::fs::File;
 mod cli;
-
-/*
-{
-  "crypto": {
-    "kdf": {
-      "function": "scrypt",
-      "params": {
-        "dklen": 32,
-        "n": 262144,
-        "r": 8,
-        "p": 1,
-        "salt": "02f43cc5370fdc150f3e613218c2473df19a8271f6cd4b3611db17ab215b6727"
-      },
-      "message": ""
-    },
-    "checksum": {
-      "function": "sha256",
-      "params": {},
-      "message": "1c55a6a94823360e26fd7b0e99d7f5342628837cbeefa89acdc6570017f6891f"
-    },
-    "cipher": {
-      "function": "aes-128-ctr",
-      "params": {
-        "iv": "454df42250b46d0d7ce5b601748c0471"
-      },
-      "message": "4ac51b3b5ca2229312f7aa8d995370ca2a72d9d1a47297611fc065650ff3043b"
-    }
-  },
-  "description": "",
-  "pubkey": "82b8b290cfd9f87071ce409ebe08f4fd70f9b486f06b2ec0b355f588a8cbb8980fb7c7af3efaf2582c425480603bcacc",
-  "path": "m/12381/3600/0/0/0",
-  "uuid": "ebc75f97-3831-4453-8b54-b8fb3763f6bf",
-  "version": 4
-}
-*/
+mod keystore;
 
 struct Operator {
     // id
@@ -48,58 +16,25 @@ struct KeyshareData {
     // operators: Vec<u32>
 }
 
-/// Representation of a keystore file generated by the deposit cli
-struct Keystore {
-    crypto: Crypto,
-    description: String,
-    pubkey: String,
-    path: String,
-    uuid: String,
-    version: u32,
-}
-
-struct Crypto {
-    kdf: Kdf,
-    checksum: Checksum,
-    cipher: Cipher,
-}
-struct Checksum {
-    function: String,
-    //params // the params
-    message: String,
-}
-
-struct Cipher {
-    function: String,
-    params: CipherParams,
-    message: String,
-}
-
-struct CipherParams {
-    iv: String,
-}
-
-struct Kdf {
-    function: String,
-    params: KdfParams,
-    message: String,
-}
-
-struct KdfParams {
-    dklen: u32,
-    n: u32,
-    r: u32,
-    p: u32,
-    salt: u32,
+pub fn start_keysplitter(keygen: Keygen) {
+    match keygen.subcommand {
+        KeygenSubcommands::Manual(manual) => manual_split(manual),
+        KeygenSubcommands::Onchain(onchain) => onchain_split(onchain),
+    }
 }
 
 // impl display for keyshare {
 // }
-fn base_processing() {
+fn base_processing(shared: &SharedKeygenOptions) {
+    let keystore_file = File::open(shared.keystore_path.clone()).unwrap();
+    let keystore = keystore::parse_keystore(keystore_file);
+
+
     // read in the keystore file
 }
 
 pub fn onchain_split(onchain: Onchain) {
+    base_processing(&onchain.shared);
     // High level steps
     // Sync all of the data from the chain
     // - should this just be in memory or should we crate keysplitting db?? probs keysplitter
@@ -107,4 +42,6 @@ pub fn onchain_split(onchain: Onchain) {
     // impl display on the key struct to easily write it into a file
 }
 
-pub fn manual_split(manual: Manual) {}
+pub fn manual_split(manual: Manual) {
+
+}
